requirements.txt
asgiref==3.8.1
axios==0.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
idna==3.10
lxml==5.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.18.0
requests==2.32.3
rich==13.8.1
sqlparse==0.5.1
urllib3==2.2.3
